🔨 Feat | 레이저 밟은 연출

+++ Assets/02. Scripts/Circus/CircusUIManager.cs
@@ -69,7 +69,7 @@ public class CircusUIManager : MonoBehaviour
         gaugeBar.value--;
     }
 
-    public void GaugeSetting(int maxValue)
+    public void GaugeSetting(float maxValue)
     {
         gaugeBar.maxValue = maxValue;
         gaugeBar.value = 0;
